Reorganize package
